+++ test/svg-sprite.test.js
@@ -16,7 +16,6 @@ const fs = require('fs');
 const path = require('path');
 const should = require('should');
 const rimraf = require('rimraf');
-const glob = require('glob');
 const looksSame = require('looks-same');
 const mustache = require('mustache');
 const sass = require('sass');
@@ -24,44 +23,18 @@ const less = require('less');
 const stylus = require('stylus');
 const SVGSpriter = require('../lib/svg-sprite.js');
 
-const compareSvg2PngHelper = require('./helpers/compare-svg-2-png.js');
 const capturePuppeteer = require('./helpers/capture-puppeteer.js');
 const writeFiles = require('./helpers/write-files.js');
 const writeFile = require('./helpers/write-file.js');
 const { addFixtureFiles } = require('./helpers/add-files.js');
 const testConfigs = require('./helpers/test-configs.js');
-const fixturesPath = require('./helpers/fixtures-path.js');
 
-const cwdAlign = path.join(fixturesPath, 'svg/css');
 const tmpPath = require('./helpers/tmp-path.js');
 const expectationsPath = require('./helpers/expectations-path.js');
 
 const dest = tmpPath;
 
-/**
- * Rasterize an SVG file and compare it to an expected image
- *
- * @param {string} svg                SVG file path
- * @param {string} png                PNG file path
- * @param {string} expected           Expected PNG file path
- * @param {string} diff               Diff file path
- * @param {Function} done             Callback
- * @param {string} msg                Message
- */
-async function compareSvg2Png(svg, png, expected, diff, done, msg) {
-    try {
-        await compareSvg2PngHelper(svg, png, expected, diff, (error, result) => {
-            should(result).ok;
-            should(error).not.ok;
-            should.ok(result.equal, msg + JSON.stringify(result.diffClusters) + png);
-            done();
-        });
-    } catch (error) {
-        console.error(error);
-        should(error).not.ok;
-        done();
-    }
-}
+const compareSvg2Png = require('./helpers/compare-svg-2-png.js');
 
 before(done => {
     rimraf(path.join(__dirname, '../tmp'), () => {
@@ -70,7 +43,6 @@ before(done => {
 });
 
 describe('svg-sprite', () => {
-    const align = glob.sync('**/*.svg', { cwd: cwdAlign });
     const previewTemplate = fs.readFileSync(path.join(__dirname, 'tmpl/css.html'), 'utf-8');
 
     // Test the minimum configuration
@@ -404,365 +376,6 @@ describe('svg-sprite', () => {
                         });
                     });
                 });
-            });
-        });
-    });
-
-    describe(`with centered alignment and ${align.length} SVG files`, () => {
-        let spriter = null;
-        let data = null;
-        const svg = {};
-
-        describe('with «css» mode, vertical layout and CSS render type', () => {
-            it('creates 2 files', done => {
-                spriter = new SVGSpriter({
-                    dest,
-                    shape: {
-                        align: path.join(fixturesPath, 'yaml/align.centered.yaml'),
-                        dimension: {
-                            maxWidth: 200,
-                            maxHeight: 200
-                        }
-                    }
-                });
-                addFixtureFiles(spriter, align, cwdAlign);
-                spriter.compile({
-                    css: {
-                        sprite: 'svg/css.vertical.centered.svg',
-                        layout: 'vertical',
-                        dimensions: true,
-                        render: {
-                            css: {
-                                dest: 'sprite.centered.css'
-                            }
-                        }
-                    }
-                }, (error, result, cssData) => {
-                    result.css.should.be.an.Object;
-                    writeFiles(result).should.be.exactly(2);
-                    data = cssData.css;
-                    svg.vertical = path.basename(result.css.sprite.path);
-                    done();
-                });
-            });
-
-            it('creates visually correct sprite', done => {
-                compareSvg2Png(
-                    path.join(tmpPath, 'css/svg', svg.vertical),
-                    path.join(tmpPath, 'css/png/css.vertical.centered.png'),
-                    path.join(expectationsPath, '/png/css.vertical.centered.png'),
-                    path.join(tmpPath, 'css/png/css.vertical.centered.diff.png'),
-                    done,
-                    'The vertical sprite doesn\'t match the expected one!'
-                );
-            });
-
-            it('creates a visually correct stylesheet resource', done => {
-                data.css = '../sprite.centered.css';
-
-                const out = mustache.render(previewTemplate, data);
-                const preview = writeFile(path.join(tmpPath, 'css/html/css.vertical.centered.html'), out);
-                const previewImage = path.join(tmpPath, 'css/png/css.vertical.centered.html.png');
-
-                preview.should.be.ok;
-
-                capturePuppeteer(preview, previewImage, error => {
-                    should(error).not.ok;
-                    looksSame(previewImage, path.join(expectationsPath, '/png/css.vertical.centered.html.png'), (error, result) => {
-                        should(error).not.ok;
-                        should.ok(result.equal, 'The generated CSS preview doesn\'t match the expected one!');
-                        done();
-                    });
-                });
-            });
-        });
-
-        describe('with «css» mode, horizontal layout and Sass render type', () => {
-            it('creates 2 files', done => {
-                spriter = new SVGSpriter({
-                    dest,
-                    shape: {
-                        align: path.join(fixturesPath, 'yaml/align.centered.yaml'),
-                        dimension: {
-                            maxWidth: 200,
-                            maxHeight: 200
-                        }
-                    }
-                });
-                addFixtureFiles(spriter, align, cwdAlign);
-                spriter.compile({
-                    css: {
-                        sprite: 'svg/css.horizontal.centered.svg',
-                        layout: 'horizontal',
-                        dimensions: true,
-                        render: {
-                            scss: {
-                                dest: 'sprite.centered.scss'
-                            }
-                        }
-                    }
-                }, (error, result, cssData) => {
-                    result.css.should.be.an.Object;
-                    writeFiles(result).should.be.exactly(2);
-                    data = cssData.css;
-                    svg.horizontal = path.basename(result.css.sprite.path);
-                    done();
-                });
-            });
-
-            it('creates visually correct sprite', done => {
-                compareSvg2Png(
-                    path.join(tmpPath, 'css/svg', svg.horizontal),
-                    path.join(tmpPath, 'css/png/css.horizontal.centered.png'),
-                    path.join(expectationsPath, '/png/css.horizontal.centered.png'),
-                    path.join(tmpPath, 'css/png/css.horizontal.centered.diff.png'),
-                    done,
-                    'The horizontal sprite doesn\'t match the expected one!'
-                );
-            });
-
-            it('creates a visually correct stylesheet resource', done => {
-                sass.render({ file: path.join(tmpPath, 'css/sprite.centered.scss') }, (err, scssText) => {
-                    should(err).not.ok;
-                    should(writeFile(path.join(tmpPath, 'css/sprite.centered.scss.css'), scssText.css)).be.ok;
-
-                    data.css = '../sprite.centered.scss.css';
-
-                    const out = mustache.render(previewTemplate, data);
-                    const preview = writeFile(path.join(tmpPath, 'css/html/scss.horizontal.centered.html'), out);
-                    const previewImage = path.join(tmpPath, 'css/png/scss.horizontal.centered.html.png');
-
-                    preview.should.be.ok;
-
-                    capturePuppeteer(preview, previewImage, error => {
-                        should(error).not.ok;
-                        looksSame(previewImage, path.join(expectationsPath, '/png/css.horizontal.centered.html.png'), (error, result) => {
-                            should(error).not.ok;
-                            should.ok(result.equal, 'The generated Sass preview doesn\'t match the expected one!');
-                            done();
-                        });
-                    });
-                }
-                );
-            });
-        });
-
-        describe('with «css» mode, packed layout and LESS render type', () => {
-            it('creates 2 files', done => {
-                spriter = new SVGSpriter({
-                    dest,
-                    shape: {
-                        align: path.join(fixturesPath, 'yaml/align.centered.yaml'),
-                        dimension: {
-                            maxWidth: 200,
-                            maxHeight: 200
-                        }
-                    }
-                });
-                addFixtureFiles(spriter, align, cwdAlign);
-                spriter.compile({
-                    css: {
-                        sprite: 'svg/css.packed.centered.svg',
-                        layout: 'packed',
-                        dimensions: true,
-                        render: {
-                            less: {
-                                dest: 'sprite.centered.less'
-                            }
-                        }
-                    }
-                }, (error, result, cssData) => {
-                    result.css.should.be.an.Object;
-                    writeFiles(result).should.be.exactly(2);
-                    data = cssData.css;
-                    svg.packed = path.basename(result.css.sprite.path);
-                    done();
-                });
-            });
-
-            it('creates visually correct sprite', done => {
-                compareSvg2Png(
-                    path.join(tmpPath, 'css/svg', svg.packed),
-                    path.join(tmpPath, 'css/png/css.packed.centered.png'),
-                    path.join(expectationsPath, '/png/css.packed.aligned.png'),
-                    path.join(tmpPath, 'css/png/css.packed.centered.diff.png'),
-                    done,
-                    'The packed sprite doesn\'t match the expected one!'
-                );
-            });
-
-            it('creates a visually correct stylesheet resource', done => {
-                const lessFile = path.join(tmpPath, 'css/sprite.centered.less');
-
-                fs.readFile(lessFile, 'utf-8', (err, lessText) => {
-                    should(err).not.ok;
-
-                    less.render(lessText, {}, (error, output) => {
-                        should(error).not.ok;
-                        should(writeFile(path.join(tmpPath, 'css/sprite.centered.less.css'), output.css)).be.ok;
-
-                        data.css = '../sprite.centered.less.css';
-
-                        const out = mustache.render(previewTemplate, data);
-                        const preview = writeFile(path.join(tmpPath, 'css/html/less.packed.centered.html'), out);
-                        const previewImage = path.join(tmpPath, 'css/png/less.packed.centered.html.png');
-
-                        preview.should.be.ok;
-
-                        capturePuppeteer(preview, previewImage, error => {
-                            should(error).not.ok;
-                            looksSame(previewImage, path.join(expectationsPath, '/png/css.packed.aligned.html.png'), (error, result) => {
-                                should(error).not.ok;
-                                should.ok(result.equal, 'The generated LESS preview doesn\'t match the expected one!');
-                                done();
-                            });
-                        });
-                    });
-                });
-            });
-        });
-    });
-
-    describe(`with mixed alignment and ${align.length} SVG files`, () => {
-        let spriter = null;
-        let data = null;
-        const svg = {};
-
-        describe('with «view» mode, vertical layout and CSS render type', () => {
-            it('creates 2 files', done => {
-                spriter = new SVGSpriter({
-                    dest,
-                    shape: {
-                        align: path.join(fixturesPath, 'yaml/align.mixed.yaml'),
-                        dimension: {
-                            maxWidth: 200,
-                            maxHeight: 200
-                        }
-                    }
-                });
-                addFixtureFiles(spriter, align, cwdAlign);
-                spriter.compile({
-                    view: {
-                        sprite: 'svg/view.vertical.mixed.svg',
-                        layout: 'vertical',
-                        dimensions: true,
-                        render: {
-                            css: {
-                                dest: 'sprite.mixed.css'
-                            }
-                        }
-                    }
-                }, (error, result, cssData) => {
-                    result.view.should.be.an.Object;
-                    writeFiles(result).should.be.exactly(2);
-                    data = cssData.view;
-                    svg.vertical = path.basename(result.view.sprite.path);
-                    done();
-                });
-            });
-
-            it('creates visually correct sprite', done => {
-                compareSvg2Png(
-                    path.join(tmpPath, 'view/svg', svg.vertical),
-                    path.join(tmpPath, 'view/png/css.vertical.mixed.png'),
-                    path.join(expectationsPath, '/png/css.vertical.mixed.png'),
-                    path.join(tmpPath, 'view/png/css.vertical.mixed.diff.png'),
-                    done,
-                    'The vertical sprite doesn\'t match the expected one!'
-                );
-            });
-
-            it('creates a visually correct stylesheet resource', done => {
-                data.css = '../sprite.mixed.css';
-
-                const out = mustache.render(previewTemplate, data);
-                const preview = writeFile(path.join(tmpPath, 'view/html/css.vertical.mixed.html'), out);
-                const previewImage = path.join(tmpPath, 'view/png/css.vertical.mixed.html.png');
-
-                preview.should.be.ok;
-
-                capturePuppeteer(preview, previewImage, error => {
-                    should(error).not.ok;
-                    looksSame(previewImage, path.join(expectationsPath, '/png/css.vertical.mixed.html.png'), (error, result) => {
-                        should(error).not.ok;
-                        should.ok(result.equal, 'The generated CSS preview doesn\'t match the expected one!');
-                        done();
-                    });
-                });
-            });
-        });
-
-        describe('with «view» mode, horizontal layout and Sass render type', () => {
-            it('creates 2 files', done => {
-                spriter = new SVGSpriter({
-                    dest,
-                    shape: {
-                        align: path.join(fixturesPath, 'yaml/align.mixed.yaml'),
-                        dimension: {
-                            maxWidth: 200,
-                            maxHeight: 200
-                        },
-                        dest: 'shapes'
-                    },
-                    svg: {
-                        namespaceIDs: true
-                    }
-                });
-                addFixtureFiles(spriter, align, cwdAlign);
-                spriter.compile({
-                    view: {
-                        sprite: 'svg/view.horizontal.mixed.svg',
-                        layout: 'horizontal',
-                        dimensions: true,
-                        render: {
-                            scss: {
-                                dest: 'sprite.mixed.scss'
-                            }
-                        }
-                    }
-                }, (error, result, cssData) => {
-                    result.view.should.be.an.Object;
-                    writeFiles(result).should.be.exactly(8);
-                    data = cssData.view;
-                    svg.horizontal = path.basename(result.view.sprite.path);
-                    done();
-                });
-            });
-
-            it('creates visually correct sprite', done => {
-                compareSvg2Png(
-                    path.join(tmpPath, 'view/svg', svg.horizontal),
-                    path.join(tmpPath, 'view/png/css.horizontal.mixed.png'),
-                    path.join(expectationsPath, '/png/css.horizontal.mixed.png'),
-                    path.join(tmpPath, 'view/png/css.horizontal.mixed.diff.png'),
-                    done,
-                    'The horizontal sprite doesn\'t match the expected one!'
-                );
-            });
-
-            it('creates a visually correct stylesheet resource', done => {
-                sass.render({ file: path.join(tmpPath, 'view/sprite.mixed.scss') }, (err, scssText) => {
-                    should(err).not.ok;
-                    should(writeFile(path.join(tmpPath, 'view/sprite.mixed.scss.css'), scssText.css)).be.ok;
-
-                    data.css = '../sprite.mixed.scss.css';
-
-                    const out = mustache.render(previewTemplate, data);
-                    const preview = writeFile(path.join(tmpPath, 'view/html/scss.horizontal.mixed.html'), out);
-                    const previewImage = path.join(tmpPath, 'view/png/scss.horizontal.mixed.html.png');
-
-                    preview.should.be.ok;
-
-                    capturePuppeteer(preview, previewImage, error => {
-                        should(error).not.ok;
-                        looksSame(previewImage, path.join(expectationsPath, '/png/css.horizontal.mixed.html.png'), (error, result) => {
-                            should(error).not.ok;
-                            should.ok(result.equal, 'The generated Sass preview doesn\'t match the expected one!');
-                            done();
-                        });
-                    });
-                }
-                );
             });
         });
     });
